requirements.txt
torch>=2.0.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
numpy>=1.21.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.28.0
aiofiles>=23.0.0
python-dotenv>=0.19.0
